+++ test/yulPhaser/SimulationRNG.cpp
@@ -95,7 +95,7 @@ BOOST_AUTO_TEST_CASE(uniformInt_returns_different_values_when_called_multiple_ti
 	constexpr double expectedValue = (minValue + maxValue) / 2.0;
 	constexpr double variance = ((maxValue - minValue + 1) * (maxValue - minValue + 1) - 1) / 12.0;
 
-	vector<uint32_t> samples;
+	vector<size_t> samples;
 	for (uint32_t i = 0; i < numSamples; ++i)
 		samples.push_back(SimulationRNG::uniformInt(minValue, maxValue));
 
@@ -110,21 +110,21 @@ BOOST_AUTO_TEST_CASE(uniformInt_can_be_reset)
 	constexpr uint32_t maxValue = 80;
 
 	SimulationRNG::reset(1);
-	vector<uint32_t> samples1;
+	vector<size_t> samples1;
 	for (uint32_t i = 0; i < numSamples; ++i)
 		samples1.push_back(SimulationRNG::uniformInt(minValue, maxValue));
 
-	vector<uint32_t> samples2;
+	vector<size_t> samples2;
 	for (uint32_t i = 0; i < numSamples; ++i)
 		samples2.push_back(SimulationRNG::uniformInt(minValue, maxValue));
 
 	SimulationRNG::reset(1);
-	vector<uint32_t> samples3;
+	vector<size_t> samples3;
 	for (uint32_t i = 0; i < numSamples; ++i)
 		samples3.push_back(SimulationRNG::uniformInt(minValue, maxValue));
 
 	SimulationRNG::reset(2);
-	vector<uint32_t> samples4;
+	vector<size_t> samples4;
 	for (uint32_t i = 0; i < numSamples; ++i)
 		samples4.push_back(SimulationRNG::uniformInt(minValue, maxValue));
 
@@ -148,7 +148,7 @@ BOOST_AUTO_TEST_CASE(binomialInt_should_produce_samples_with_right_expected_valu
 	constexpr double expectedValue = numTrials * successProbability;
 	constexpr double variance = numTrials * successProbability * (1 - successProbability);
 
-	vector<uint32_t> samples;
+	vector<size_t> samples;
 	for (uint32_t i = 0; i < numSamples; ++i)
 		samples.push_back(SimulationRNG::binomialInt(numTrials, successProbability));
 
@@ -163,21 +163,21 @@ BOOST_AUTO_TEST_CASE(binomialInt_can_be_reset)
 	constexpr double successProbability = 0.6;
 
 	SimulationRNG::reset(1);
-	vector<uint32_t> samples1;
+	vector<size_t> samples1;
 	for (uint32_t i = 0; i < numSamples; ++i)
 		samples1.push_back(SimulationRNG::binomialInt(numTrials, successProbability));
 
-	vector<uint32_t> samples2;
+	vector<size_t> samples2;
 	for (uint32_t i = 0; i < numSamples; ++i)
 		samples2.push_back(SimulationRNG::binomialInt(numTrials, successProbability));
 
 	SimulationRNG::reset(1);
-	vector<uint32_t> samples3;
+	vector<size_t> samples3;
 	for (uint32_t i = 0; i < numSamples; ++i)
 		samples3.push_back(SimulationRNG::binomialInt(numTrials, successProbability));
 
 	SimulationRNG::reset(2);
-	vector<uint32_t> samples4;
+	vector<size_t> samples4;
 	for (uint32_t i = 0; i < numSamples; ++i)
 		samples4.push_back(SimulationRNG::binomialInt(numTrials, successProbability));
